Extract element discussion proxy as mixin
- Handled  store state migration within element bus in proxy mixin above
